Initial FFT stuff

+++ led.h
@@ -1,8 +1,10 @@
 #ifndef _LED_H_
 #define _LED_H_
 
+#include "defs.h"
+
 void led_init();
-void led_set1(int r, int g, int b);
-void led_set2(int r, int g, int b);
+void led_set1(u8 r, u8 g, u8 b);
+void led_set2(u8 r, u8 g, u8 b);
 
 #endif
